refactor: 레스토랑 관련 컴포넌트 분리<detail>
