+++ gonggibap/src/app/_components/sidebar/restaurant/detail/index.ts
@@ -1,0 +1,3 @@
+export { RestaurantDetailView } from "@/app/_components/sidebar/restaurant/detail/RestaurantDetailView";
+export { RestaurantHeader } from "@/app/_components/sidebar/restaurant/detail/RestaurantHeader";
+export { RestaurantInfo } from "@/app/_components/sidebar/restaurant/detail/RestaurantInfo";
